Lesson7 mailroom. This is really WIP.
